% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance_rate.R
\name{acceptance_rate}
\alias{acceptance_rate}
\title{Compute acceptance rate}
\usage{
acceptance_rate(...)
}
\arguments{
\item{...}{parameters to \code{\link{get_traces}} (especially 'run')}
}
\value{
acceptance rate
}
\description{
This function takes the provided \code{\link{libbi}} object which has been
run, or a bi file, and returns a the acceptance rate
}
\examples{
example_run_file <- system.file(package="rbi.helpers", "example_run.nc")
example_model_file <- system.file(package="rbi", "PZ.bi")
example_bi <- attach_file(libbi(example_model_file), "output", example_run_file)
acceptance_rate(example_bi)
}
\author{
Sebastian Funk
}
