% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_particles.R
\name{adapt_particles}
\alias{adapt_particles}
\title{Adapt the number of particles}
\usage{
adapt_particles(wrapper, min = 1, max = 1024, add_options, samples, ...)
}
\arguments{
\item{wrapper}{\code{\link{libbi}} (which has been run) to study}

\item{min}{minimum number of particles}

\item{max}{maximum number of particles}

\item{add_options}{list of additional options}

\item{samples}{number of samples to generate each iteration}

\item{...}{parameters for libbi$run}
}
\value{
a \code{\link{libbi}} with the desired proposal distribution
}
\description{
This function takes the provided \code{\link{libbi}} and
  runs MCMC at a single point (i.e., repeatedly proposing the same paramters),
  adapting the number of particles distribution until the variance of the likelihood
  crosses one.
}

