% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_proposal.R
\name{adapt_proposal}
\alias{adapt_proposal}
\title{Adapt the proposal distribution of MCMC using the covariance
  of samples}
\usage{
adapt_proposal(wrapper, min = 0, max = 1, scale = 2, options, nsamples,
  max_iter = 10, correlations = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{wrapper}{\code{link{libbi}} (which has been run) to study}

\item{min}{minimum acceptance rate}

\item{max}{maximum acceptance rate}

\item{scale}{scale multiplier/divider for the proposal. If >1 this
will be inverted.}

\item{options}{list of additional options}

\item{nsamples}{number of samples to generate each iteration}

\item{max_iter}{maximum of iterations (default: 10)}

\item{correlations}{if TRUE (default), take into account correlations}

\item{quiet}{if set to TRUE, will not provide running output of particle numbers tested}

\item{...}{parameters for \code{\link{sample}}}
}
\value{
a \code{\link{libbi}} with the desired proposal distribution
}
\description{
This function takes the provided \code{\link{libbi}} object and
  runs MCMC, adapting the proposal distribution until the desired
  acceptance rate is achieved. If a scale is given, it will be used
  to adapt the proposal at each iteration
}
\examples{
example_obs <- bi_read(system.file(package="rbi", "example_output.nc"))
example_model <- bi_model(system.file(package="rbi", "PZ.bi"))
example_bi <- libbi(model = example_model, obs = example_obs)
obs_states <- var_names(example_model, "obs")
max_time <- max(sapply(example_obs[obs_states], function(x) { max(x[["time"]])}))
# adapt to acceptance rate between 0.1 and 0.5
\dontrun{adapted <- adapt_proposal(example_bi, nsamples = 100, end_time = max_time,
                               min = 0.1, max = 0.5, nparticles = 256, correlations = TRUE)}
}

