% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_DIC.R
\name{compute_DIC}
\alias{compute_DIC}
\title{Compute DIC for a libbi model}
\usage{
compute_DIC(read, burn)
}
\arguments{
\item{read}{either a \code{libbi} object or a list of data frames, as returned by \code{bi_read}}

\item{burn}{number of iterations to discard as burn-in (if any)}
}
\value{
DIC
}
\description{
Compute DIC for a libbi model
}
\author{
Sebastian Funk
}

