% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorise_columns.R
\name{factorise_columns}
\alias{factorise_columns}
\title{Factorise columns in a data frame}
\usage{
factorise_columns(df, labels)
}
\arguments{
\item{df}{The data frame (or data table) to manipulate}

\item{labels}{(optionally), a character vector of new labels for the table elements; the names of the character vector should be (some or all of) the old elements}
}
\value{
a data.table with updated columns
}
\description{
Optionally, can give new labels
}
\keyword{internal}

