% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mvn_params.r
\name{get_mvn_params}
\alias{get_mvn_params}
\title{Construct a covariance matrix}
\usage{
get_mvn_params(x, scale = 1, correlations = TRUE, fix)
}
\arguments{
\item{x}{a \code{\link{libbi}} which has been run}

\item{scale}{a factor by which to scale all elements of the covariance matrix}

\item{correlations}{logical; if TRUE, correlations are taken into account
when constructing the parameters}

\item{fix}{if this is set, all elements of the covariance matrix will be set
to it}
}
\value{
the updated bi model
}
\description{
This function takes the provided \code{\link{libbi}} which has been
run and returns the square root of the covariance matrix, which
can be used for proposal distributions
}
\keyword{internal}
