% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.r
\name{numeric_to_time}
\alias{numeric_to_time}
\title{Convert numeric times to actual times or dates}
\usage{
numeric_to_time(x, origin, unit, ...)
}
\arguments{
\item{x}{a [rbi::libbi()] object which has been run, or a list of data
frames containing state trajectories (as returned by \code{bi_read})}

\item{origin}{the time origin, i.e. the date or time corresponding to time 0}

\item{unit}{the unit of time that each time step corresponds to; this must be
a unit understood by \code{lubridate::period}, optionally with a number in
advance, e.g. "day" or "2 weeks" or "3 seconds"}

\item{...}{any arguments for \code{bi_read} (e.g., \code{file})}
}
\value{
a list of data frames as returned by \code{bi_read}, but with real
  times
}
\description{
This function converts from numeric times (i.e., 0, 1, 2, ...)
  to actual times or dates
}
