% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_to_cov.R
\name{output_to_cov}
\alias{output_to_cov}
\title{Construct a covariance matrix}
\usage{
output_to_cov(x, scale, correlations = FALSE)
}
\arguments{
\item{x}{a \code{\link{libbi}} which has been run}

\item{scale}{a factor by which to scale all the standard deviations}

\item{correlations}{whether to take into account correlations}
}
\value{
the updated bi model
}
\description{
This function takes the provided \code{\link{libbi}} which has been
run and returns the square root of the covariance matrix, which
can be used for proposal distributions
}
\keyword{internal}
