% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_to_proposal.R
\name{output_to_proposal}
\alias{output_to_proposal}
\title{Construct a proposal from run results}
\usage{
output_to_proposal(wrapper, scale, correlations = FALSE,
  truncate = TRUE, blocks = c("parameter", "initial"))
}
\arguments{
\item{wrapper}{a \code{\link{libbi}} which has been run}

\item{scale}{a factor by which to scale all the standard deviations}

\item{correlations}{whether to take into account correlations}

\item{truncate}{truncate the multivariate normal proposals according to the used priors, e.g. truncating a parameter with beta prior at 0 and 1}

\item{blocks}{blocks to use (out of "parameter" and "initial")}
}
\value{
the updated bi model
}
\description{
This function takes the provided \code{\link{libbi}} which has been
run and returns a new model which has the proposal constructed from
the sample mean and standard deviation.
}
\keyword{internal}
