% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_libbi.R
\name{plot.libbi}
\alias{plot.libbi}
\title{Plot routing for \code{libbi} objects}
\usage{
\method{plot}{libbi}(obj, ...)
}
\arguments{
\item{obj}{\code{libbi} object}

\item{...}{parameters to \code{\link{plot_libbi}}}
}
\value{
plot
}
\description{
Plot routing for \code{libbi} objects
}

