% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_libbi.R
\name{plot_libbi}
\alias{plot_libbi}
\title{Plot results from libbi}
\usage{
plot_libbi(data, model, prior, states, params, noises, quantiles = c(0.5,
  0.95), date.origin, date.unit, time.dim = "time", obs, id, extra.aes,
  all.times = FALSE, hline, burn, steps = FALSE, select, shift,
  obs.colour = "red", base.alpha = 0.5, trend = "median",
  densities = "histogram", density_args = list(), limit.to.obs = FALSE,
  labels, brewer.palette, plot = TRUE, ...)
}
\arguments{
\item{data}{Monte-Carlo samples, either a \code{libbi} object or a list of data frames, as returned by \code{bi_read}, or the name of an NetCDF file used as 'output' in a libbi run}

\item{model}{model file or a \code{bi_model} object (if \code{data} is not a \code{libbi} object)}

\item{prior}{optional; Prior samples, either a \code{libbi} object or a list of data frames, as returned by \code{bi_read}}

\item{states}{states to plot (if not given, all states will be plotted; if empty vector is passed, no states are plotted)}

\item{params}{parameters to plot (if not given, all states will be plotted; if empty vector is passed, no parameters are plotted)}

\item{noises}{noises to plot (if not given, all noises will be plotted; if empty vector is passed, no noises are plotted)}

\item{quantiles}{if plots are produced, which quantile to use for confidence intervals (NULL for no confidence intervals)}

\item{date.origin}{date of origin (if dates are to be calculated)}

\item{date.unit}{unit of date (if desired, otherwise the time dimension will be used)}

\item{time.dim}{time dimension ("time" by default)}

\item{obs}{observations (a named list of data frames, a \code{libbi} object with observations, or a NetCDF file name)}

\item{id}{one or more run ids to plot}

\item{extra.aes}{extra aesthetics (for ggplot)}

\item{all.times}{whether to plot all times (not only ones with observations)}

\item{hline}{horizontal marker lines, named vector in format (state = value)}

\item{burn}{How many iterations to burn}

\item{steps}{whether to plot lines as stepped lines}

\item{select}{list of selection criteria}

\item{shift}{list of dimensions to be shifted, and by how much}

\item{obs.colour}{colour for plotting the observations}

\item{base.alpha}{base alpha value for credible intervals}

\item{trend}{how the trend should be characterised (e.g., mean, median, or NULL for no trend line)}

\item{densities}{density geometry (e.g., "histogram" (default) or "density")}

\item{density_args}{list of arguments to pass to density geometry}

\item{limit.to.obs}{whether to limit the time axis to times with observations}

\item{labels}{facet labels, in case they are to be rewritten, to be parsed using \code{label_parsed}; should be given as named character vector of (parameter = 'label') pairs}

\item{brewer.palette}{optional; brewer color palette}

\item{plot}{set to FALSE to suppress plot of trajectories}

\item{...}{options for geom_step / geom_line / geom_point / etc.}
}
\value{
a list of plots: states, densities, traces, correlations, noises, likelihoods, as well as underlying raw and aggregate data
}
\description{
Plots state trajectories (unless plot = FALSE) and invisibly returns a list of state trajectories and other plots.
}
\examples{
example_run_file <- system.file(package="rbi.helpers", "example_run.nc")
example_model_file <- system.file(package="rbi", "PZ.bi")
example_bi <- recreate_libbi(example_model_file, example_run_file)

plot(example_bi) # just plot trajectories
p <- plot(example_bi, plot = FALSE) # get whole suite of plots

p$states
p$densities
p$traces
p$correlations
p$noises
p$likelihoods
}
\author{
Sebastian Funk
}

