% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.r
\name{real_to_time}
\alias{real_to_time}
\title{Convert real times to LibBi times}
\usage{
real_to_time(x, origin, unit)
}
\arguments{
\item{x}{a data frames containing a "time" column, or a list containing such data frames}

\item{origin}{the time origin, i.e. the date or time corresponding to LibBi
time 0}

\item{unit}{the unit of time that each time step corresponds to; this must bg
a unit understood by \code{lubridate::period}, optionally with a number in
advance, e.g. "day" or "2 weeks" or "3 seconds"}
}
\value{
a list of data frames that can be passed to libbi
}
\description{
This function converts from real times/dates to LibBi times
}
