% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_libbi.R
\name{recreate_libbi}
\alias{recreate_libbi}
\title{Recreate libbi wrapper}
\usage{
recreate_libbi(model, output_file_name, client = "sample", ...)
}
\arguments{
\item{model}{either a \code{\link{bi_model}} object, or a path to a model file}

\item{output_file_name}{the NetCDF file holding the output of the \code{libbi} run}

\item{client}{defaults to "sample"; this won't be needed in most cases but prevents a warning.}

\item{...}{further options to pass to \code{libbi}, e.g. global_options}
}
\value{
\code{\link{libbi}} object that can be analysed
}
\description{
This function links a model and output file name of a previous run and
returns a \code{\link{libbi}} object, so as to be able to pass it to analysis
functions
}
\examples{
example_run_file <- system.file(package="rbi.helpers", "example_run.nc")
example_model_file <- system.file(package="rbi", "PZ.bi")
example_bi <- recreate_libbi(example_model_file, example_run_file)
}

