% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.r
\name{split_unit}
\alias{split_unit}
\title{Split a unit string}
\usage{
split_unit(unit_string)
}
\arguments{
\item{unit_string}{the string to split}
}
\value{
a list with two elements, "num" (the amount) and "unit", , for use with lubridate::period
}
\description{
Splits a unit string (e.g., "2 weeks") into the amount (2) and unit ("weeks")
}
\author{
Sebastian Funk
}
\keyword{internal}
