% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.r
\name{time_to_numeric}
\alias{time_to_numeric}
\title{Convert actual times or dates to numeric times}
\usage{
time_to_numeric(x, origin, unit)
}
\arguments{
\item{x}{a data frame containing a "time" column, or a list containing such
data frames}

\item{origin}{the time origin, i.e. the date or time corresponding to time 0}

\item{unit}{the unit of time that each time step corresponds to; this must be
a unit understood by \code{lubridate::period}, optionally with a number in
advance, e.g. "day" or "2 weeks" or "3 seconds"}
}
\value{
a list of data frames that can be passed to \code{libbi}
}
\description{
This function converts from real times/dates to numeric times
  (0, 1, 2, ...)
}
