% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.r
\name{to_actual_times}
\alias{to_actual_times}
\title{Convert LibBi times to actual times or dates}
\usage{
to_actual_times(x, origin, unit, ...)
}
\arguments{
\item{x}{a \code{\link{libbi}} object which has been run, or a list of data
frames containing state trajectories (as returned by \code{bi_read})}

\item{origin}{the time origin, i.e. the date or time corresponding to LibBi
time 0}

\item{unit}{the unit of time that each time step corresponds to; this must bg
a unit understood by \code{lubridate::period}, optionally with a number in
advance, e.g. "day" or "2 weeks" or "3 seconds"}

\item{...}{any arguments for \code{bi_read} (e.g., \code{file})}
}
\value{
a list of data frames as returned bi \code{bi_read}, but with real times
}
\description{
This function converts from LibBi times (i.e., 0, 1, 2, ...) to actual times or dates
}
