% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_proposal.r
\name{update_proposal}
\alias{update_proposal}
\title{Construct a proposal from run results}
\usage{
update_proposal(model, truncate = TRUE, blocks = c("parameter", "initial"))
}
\arguments{
\item{model}{a [rbi::bi_model()] object}

\item{truncate}{truncate the multivariate normal proposals according to the
used priors, e.g. truncating a parameter with beta prior at 0 and 1}

\item{blocks}{blocks to use (out of "parameter" and "initial")}
}
\value{
the updated bi model
}
\description{
This function takes the provided [rbi::bi_model()] and adds a generic
Gaussian proposal distribution.
}
\keyword{internal}
