% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance_rate.R
\name{acceptance_rate}
\alias{acceptance_rate}
\title{Compute acceptance rate}
\usage{
acceptance_rate(...)
}
\arguments{
\item{...}{parameters to [rbi::get_traces()] (especially 'x')}
}
\value{
acceptance rate
}
\description{
This function takes the provided [rbi::libbi()] object which has been
run, or a bi file, and returns a the acceptance rate
}
\examples{
example_run <- rbi::bi_read(
  system.file(package = "rbi.helpers", "example_run.nc")
)
example_model_file <- system.file(package = "rbi", "PZ.bi")
example_bi <- rbi::attach_data(
  rbi::libbi(example_model_file), "output", example_run
)
acceptance_rate(example_bi)
}
\author{
Sebastian Funk
}
