% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adapt_particles.R
\name{adapt_particles}
\alias{adapt_particles}
\title{Adapt the number of particles}
\usage{
adapt_particles(wrapper, test = 2^(0:10), add_options, samples, ...)
}
\arguments{
\item{wrapper}{\code{\link{libbi}} (which has been run) to study}

\item{add_options}{list of additional options}

\item{samples}{number of samples to generate each iteration}

\item{...}{parameters for libbi$run}
}
\value{
a \code{\link{libbi}} with the desired proposal distribution
}
\description{
This function takes the provided \code{\link{libbi}} and
  runs MCMC at a single point (i.e., repeatedly proposing the same paramters),
  adapting the number of particles distribution until the desired
  acceptance rate is achieved. If a scale is given, it will be used
  to adapt the number of particles at each iteration.
}

