% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt_particles.R
\name{adapt_particles}
\alias{adapt_particles}
\title{Adapt the number of particles}
\usage{
adapt_particles(wrapper, min = 1, max = 1024, add_options, samples, ...)
}
\arguments{
\item{wrapper}{\code{\link{libbi}} (which has been run) to study}

\item{min}{minimum number of particles}

\item{max}{maximum number of particles}

\item{add_options}{list of additional options}

\item{samples}{number of samples to generate each iteration}

\item{...}{parameters for libbi$run}
}
\value{
a \code{\link{libbi}} with the desired proposal distribution
}
\description{
This function takes the provided \code{\link{libbi}} and
  runs MCMC at a single point (i.e., repeatedly proposing the same paramters),
  adapting the number of particles distribution until the variance of the likelihood
  crosses one.
}
\examples{
example_obs <- bi_read(system.file(package="rbi", "example_output.nc"))
example_model <- bi_model(system.file(package="rbi", "PZ.bi"))
example_bi <- libbi(model = example_model, obs = example_obs)
obs_states <- example_model$get_vars("obs")
max_time <- max(sapply(example_obs[obs_states], function(x) { max(x[["time"]])}))
\dontrun{adapted <- adapt_particles(example_bi, samples = 128, end_time = max_time)}
}

