% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_DIC.R
\name{compute_DIC}
\alias{compute_DIC}
\title{Compute DIC for a libbi model}
\usage{
compute_DIC(read, burn, bootstrap = 0)
}
\arguments{
\item{read}{either a \code{libbi} object or a list of data frames, as returned by \code{bi_read}}

\item{burn}{number of iterations to discard as burn-in (if any)}

\item{bootstrap}{number of bootstrap samples to take, 0 to just take data}
}
\value{
DIC
}
\description{
Compute DIC for a libbi model
}
\examples{
example_run_file <- system.file(package="rbi", "example_output.nc")
example_model_file <- system.file(package="rbi", "PZ.bi")
example_bi <- add_output(libbi(example_model_file), example_run_file)
compute_DIC(example_bi)
}
\author{
Sebastian Funk
}
