% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_libbi.R
\name{plot.libbi}
\alias{plot.libbi}
\title{Plot routing for \code{libbi} objects}
\usage{
\method{plot}{libbi}(x, ...)
}
\arguments{
\item{x}{\code{libbi} object}

\item{...}{parameters to \code{\link{plot_libbi}}}
}
\value{
a list of plots plot (see \code{\link{plot_libbi}})
}
\description{
Plot routing for \code{libbi} objects
}
