% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_libbi.R
\name{plot_libbi}
\alias{plot_libbi}
\title{Plot results from libbi}
\usage{
plot_libbi(read, model, states, params, noises, quantile.span = c(0.5, 0.95),
  date.origin, date.unit, time.dim = "nr", data, id, extra.aes = c(),
  all.times = FALSE, hline, burn, thin, steps = FALSE, select, shift,
  data.colour = "red", base.alpha = 0.5, trend = "median",
  densities = "density", density_args = NULL, limit.to.data = FALSE, ...)
}
\arguments{
\item{read}{either a \code{libbi} object or a list of data frames, as returned by \code{bi_read}}

\item{model}{model file or a \code{bi_model} object (if \code{read} is not a \code{libbi} object)}

\item{states}{states to plot (if not given, all states will be plotted; if empty vector is passed, no states are plotted)}

\item{params}{parameters to plot (if not given, all states will be plotted; if empty vector is passed, no parameters are plotted)}

\item{noises}{noises to plot (if not given, all noises will be plotted; if empty vector is passed, no noises are plotted)}

\item{quantile.span}{if plots are produced, which quantile to use for confidence intervals}

\item{date.origin}{date of origin (if dates are to be calculated)}

\item{date.unit}{unit of date (if desired, otherwise the time dimension will be used)}

\item{time.dim}{time dimension ("nr" by default)}

\item{data}{data (with a "time" and "value" column)}

\item{id}{one or more run ids to plot}

\item{extra.aes}{extra aesthetics (for ggplot)}

\item{all.times}{whether to plot all times (not only ones with data)}

\item{hline}{horizontal marker lines, named vector in format (state = value)}

\item{burn}{How many runs to burn}

\item{thin}{How many runs to thin per run kept}

\item{steps}{whether to plot lines as stepped lines}

\item{select}{list of selection criteria}

\item{shift}{list of dimensions to be shifted, and by how much}

\item{data.colour}{colour for plotting the data}

\item{base.alpha}{base alpha value for credible intervals}

\item{trend}{how the trend should be characterised (e.g., mean, median)}

\item{densities}{density geometry (e.g., "histogram")}

\item{density_args}{list of arguments to pass to density geometry}

\item{limit.to.data}{whether to limit the time axis to times in the data}

\item{...}{options for geom_step / geom_line}
}
\value{
list of results
}
\description{
Plot results from libbi
}
\author{
Sebastian Funk
}

