% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_libbi.R
\name{plot_libbi}
\alias{plot_libbi}
\title{Plot results from libbi}
\usage{
plot_libbi(x, model, prior, type = c("state", "noise", "obs", "param",
  "logeval"), quantiles = c(0.5, 0.95), date.origin, date.unit,
  time.dim = "time", data, extra.aes, all.times = FALSE, hline, burn,
  steps = FALSE, select, data.colour = "red", base.alpha = 0.5,
  np.alpha = 0.35, trend = "median", densities = "histogram",
  density_args = list(), limit.to.data = FALSE, labels, brewer.palette,
  pairs = TRUE, correlations = TRUE, verbose = FALSE, plot = TRUE, ...)
}
\arguments{
\item{x}{Monte-Carlo samples, either a \code{libbi} object or a list of data frames, as returned by \code{bi_read}, or the name of an NetCDF file used as 'output' in a libbi run}

\item{model}{model file or a \code{bi_model} object (if \code{data} is not a \code{libbi} object)}

\item{prior}{optional; Prior samples, either a \code{libbi} object or a list of data frames, as returned by \code{bi_read}}

\item{type}{character vector determining which plots to generate; options are: "state", "obs", "param", "noise", "logevals"; by default, all will be plotted; more specific selections of variables can be given as arguments with the name of the type containing character vectors of variables, e.g. \code{param="alpha"} to just plot parameter alpha (requiring "param" to be given as one of "type")}

\item{quantiles}{if plots are produced, which quantile to use for confidence intervals (NULL for no confidence intervals)}

\item{date.origin}{date of origin (if dates are to be calculated)}

\item{date.unit}{unit of date (if desired, otherwise the time dimension will be used)}

\item{time.dim}{time dimension ("time" by default)}

\item{data}{observations (a named list of data frames, a \code{libbi} object with observations, or a NetCDF file name)}

\item{extra.aes}{extra aesthetics (for ggplot)}

\item{all.times}{whether to plot all times (not only ones with observations)}

\item{hline}{horizontal marker lines, named vector in format (state = value)}

\item{burn}{How many iterations to burn}

\item{steps}{whether to plot lines as stepped lines}

\item{select}{list of selection criteria, as named list of selected elements. If the list contains "np", it is treated specially.}

\item{data.colour}{colour for plotting the data}

\item{base.alpha}{base alpha value for credible intervals}

\item{np.alpha}{alpha of trajectories, if 'np' is part of \code{select} (default: 0.35)}

\item{trend}{how the trend should be characterised (e.g., mean, median, or NULL for no trend line)}

\item{densities}{density geometry (e.g., "histogram" (default) or "density")}

\item{density_args}{list of arguments to pass to density geometry}

\item{limit.to.data}{whether to limit the time axis to times with observations (default: FALSE)}

\item{labels}{facet labels, in case they are to be rewritten, to be parsed using \code{label_parsed}; should be given as named character vector of (parameter = 'label') pairs}

\item{brewer.palette}{optional; brewer color palette}

\item{pairs}{logical: whether to generate a pair plot}

\item{correlations}{logical: whether to generate a correlations plot}

\item{verbose}{if set to TRUE, additional output will be displayed}

\item{plot}{set to FALSE to suppress plot of trajectories}

\item{...}{more specific selection of variables to plot (see the \code{type} option); any other options will be interpreted as options for geom_step / geom_line / geom_point / etc. when plotting states/noises/observations, e.g. lwd or others}
}
\value{
a list of plots: states, densities, traces, correlations, noises, logdensities, as well as underlying raw and aggregate data
}
\description{
Plots state trajectories (unless plot = FALSE) and invisibly returns a list of state trajectories and other plots.
}
\examples{
example_run_file <- system.file(package="rbi", "example_output.nc")
example_model_file <- system.file(package="rbi", "PZ.bi")
example_bi <- add_output(libbi(example_model_file), example_run_file)

plot(example_bi) # just plot trajectories
p <- plot(example_bi, plot = FALSE) # get whole suite of plots

p$trajectories
p$correlations
p$pairs
p$densities
p$traces
p$logevals
}
\author{
Sebastian Funk
}

